Initial commit. Squashed from commit 0d2428f9d393de0bb8c8b79a96d86803eb3eba40 from TT BBE main
Notice regarding the squashed initial commit:
As we transition from our internal development environment to a public
repository, we are committed to ensuring the security and
confidentiality of sensitive information.

To protect against any accidental release of confidential data, we have
made a specific decision regarding the initial commit to our public
repository. We have opted for a squashed commit for the initial public
release, which consolidates all the changes from our internal history
into a single, clean commit. This step ensures that no sensitive
information from our past development is exposed to the public.

We want to emphasize that moving forward, we are fully dedicated to
maintaining a transparent and organized version history for all our
future commits and bug fixes. This approach allows us to maintain a
clean slate for our public repository while ensuring that our ongoing
work is well-documented and traceable.

We appreciate your understanding and support as we make this transition.
Our goal is to provide you with a secure and reliable open-source
project. If you have any questions or concerns, please don't hesitate to
reach out to us.

Thank you for your continued interest and participation in our project.