Enable fusing of non colum matmul part 2
* Enable fusing of non colum matmul part 2

In change #eda15793 it was enabled to fuse mamtul where input1 is not a
column. Limitation was that both inputs of the matmul have to be op
inputs. This change partially relaxes that limitation. Input0 can be an
intermed buffer. Cost of this solution is that intermed buffer size has
to be extended to accomodate one row of ublocks for input0 (m_k).

Internal fused_op broadcast for input0 of matmul are not tested yet.

fix typo